add long insert exonic read remover
